test: add happy GroupBy tests * restructure grouping tests into own file * fix GroupId precedence of i64 vs u64
